+++ packages/karrotframe/package.json
@@ -1,6 +1,6 @@
 {
   "name": "karrotframe",
-  "version": "0.16.0-alpha.6",
+  "version": "0.16.0-alpha.7",
   "description": "UX Enhancer",
   "repository": "https://github.com/daangn/karrotframe",
   "license": "Apache-2.0",
